pyproject.toml
[project]
name = "email-trigger"
version = "1.0.0"
description = "Send templated emails to your warden via Gmail"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "motor==3.6.0",
    "python-dotenv==1.0.0",
    "google-auth==2.27.0",
    "google-auth-oauthlib==1.2.0",
    "google-api-python-client==2.116.0",
    "python-multipart==0.0.6",
    "pydantic[email]==2.6.0",
    "pydantic-settings==2.1.0",
    "cryptography==42.0.2",
    "httpx==0.26.0",
    "itsdangerous==2.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "mongomock-motor==0.0.29",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
